requirements.txt
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
daemon==1.2
idna==2.6
kafka-python==1.4.3
msgpack-python==0.5.6
pycodestyle==2.4.0
PyYAML==3.12
requests==2.18.4
urllib3==1.22
